Addressed #333. Allow JSON to be loaded from any stream, including memory.
There was a request (#333) to be able to load Lottie files from memory. This change allows that by enabling any IInputStream to be used.

As part of this change, I found some cruft that I could clean up, and I pulled out the classes for the Loader and ContentFactory and did some other straightforward refactoring.
